Rename "home" page to "hello" in API
